add functions to utils
