+++ front-end/src/utils/map.js
@@ -58,3 +58,18 @@ export function formatInfoWindow(stoop) {
         </div>
     `
 }
+
+export function calculateDistance(lat1, lon1, lat2, lon2) {
+	const R = 3958.8 // Earth's radius in miles
+	const dLat = (lat2 - lat1) * (Math.PI / 180)
+	const dLon = (lon2 - lon1) * (Math.PI / 180)
+	const a =
+		Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+		Math.cos(lat1 * (Math.PI / 180)) *
+			Math.cos(lat2 * (Math.PI / 180)) *
+			Math.sin(dLon / 2) *
+			Math.sin(dLon / 2)
+	const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a))
+	const distance = R * c
+	return distance
+}
